add: Toast追加

+++ src/components/todoForm/hooks/useDeleteTodo.ts
@@ -1,31 +1,40 @@
 import { useState } from 'react';
 import useFirebaseApi from '../../../api/useFirebaseApi';
-import { useCookiesHooks, useToast } from '../../../hooks';
+import { useToast } from '../../../hooks';
 import { ResponseTodoType, TodoPropsType } from '../../../types';
+import useTodoValidation from './useTodoValidation';
 
 const useDeleteTodo = ({ todo, onAddTodo, onDeleteTodo }: TodoPropsType) => {
 
-  const { updateSessionTime } = useCookiesHooks();
   const { deleteTodo } = useFirebaseApi();
   const [stockTask, setStockTask] = useState(todo);
   const toast = useToast();
-
+  const validation = useTodoValidation();
+  const { cookies } = validation;
 
   const handleDelete = async () => {
-    updateSessionTime();
+    if (cookies.uid == null) {
+      return;
+    }
+    validation.sessionCheck();
     setStockTask(todo);
     onDeleteTodo(todo.id);
     const id = toast.loadingToast();
-    const response: ResponseTodoType = await deleteTodo({ uid: todo.uid, id: todo.id });
-    if (response.statusCode === 200) {
-      return toast.successToast(id);
+    try {
+      const response: ResponseTodoType = await deleteTodo({ uid: todo.uid, id: todo.id });
+      if (response.statusCode === 200) {
+        return toast.successToast(id);
+      }
+      if (response.statusCode === 204) {
+        return toast.successToast(id);
+      }
+      onAddTodo(stockTask);
+      toast.errorToast(id);
+      return;
+    } catch (error) {
+      onAddTodo(stockTask);
+      toast.errorToast(id);
     }
-    if (response.statusCode === 204) {
-      return toast.successToast(id);
-    }
-    onAddTodo(stockTask);
-    toast.errorToast(id);
-    return;
   };
 
   return { handleDelete };
